requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
torch>=2.0.0
numpy<2.0
pyyaml
python-multipart
torchtext==0.17.2
torchvision==0.17.2
openai-whisper